backend/requirements.txt
# Cognitive Agent Platform — Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# AI / LLM
google-generativeai==0.8.4
huggingface-hub==0.27.1
openai==1.59.7

# Vector Database
pinecone-client==5.0.1

# Agent Orchestration
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langgraph==0.2.62

# NLP & Data
tiktoken==0.8.0
beautifulsoup4==4.12.3
httpx==0.28.1
aiofiles==24.1.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7

# Utilities
tenacity==9.0.0
structlog==24.4.0
